[no-issue]: Extract mergeAnalytics to pure function
